(refactor) step02 리뷰 반영
